Polish: select last block if empty default block
