+++ packages/editor/src/components/editor-canvas/index.js
@@ -17,10 +17,10 @@ import {
 	__experimentalUseResizeCanvas as useResizeCanvas,
 } from '@wordpress/block-editor';
 import { useEffect, useRef, useMemo } from '@wordpress/element';
-import { useRegistry, useSelect } from '@wordpress/data';
+import { useSelect } from '@wordpress/data';
 import { parse } from '@wordpress/blocks';
 import { store as coreStore } from '@wordpress/core-data';
-import { useMergeRefs, useRefEffect } from '@wordpress/compose';
+import { useMergeRefs } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -30,6 +30,7 @@ import { store as editorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 import EditTemplateBlocksNotification from './edit-template-blocks-notification';
 import useSelectNearestEditableBlock from '../../hooks/use-select-nearest-editable-block';
+import { usePaddingAppender } from './use-padding-appender';
 
 const {
 	LayoutStyle,
@@ -95,7 +96,6 @@ function EditorCanvas( {
 	iframeProps,
 	children,
 } ) {
-	const registry = useRegistry();
 	const {
 		renderingMode,
 		postContentAttributes,
@@ -307,35 +307,7 @@ function EditorCanvas( {
 
 	const localRef = useRef();
 	const typewriterRef = useTypewriter();
-	const paddingAppenderRef = useRefEffect( ( node ) => {
-		function onMouseDown( event ) {
-			if ( event.target !== node ) {
-				return;
-			}
-
-			// only handle clicks under the last child
-			const lastChild = node.lastElementChild;
-			if ( ! lastChild ) {
-				return;
-			}
-
-			const lastChildRect = lastChild.getBoundingClientRect();
-			if ( event.clientY < lastChildRect.bottom ) {
-				return;
-			}
-
-			const { insertDefaultBlock } =
-				registry.dispatch( blockEditorStore );
-
-			insertDefaultBlock();
-
-			event.preventDefault();
-		}
-		node.addEventListener( 'mousedown', onMouseDown );
-		return () => {
-			node.removeEventListener( 'mousedown', onMouseDown );
-		};
-	}, [] );
+	const paddingAppenderRef = usePaddingAppender();
 	const contentRef = useMergeRefs( [
 		localRef,
 		renderingMode === 'post-only' ? typewriterRef : null,
